+++ week_1/1_helloworld.rb
@@ -1,0 +1,5 @@
+class HelloWorld
+    def say_hello
+        "Hello, world!"
+    end
+end
