+++ week_1/1_pairs.rb
