+++ week_1/1_ruby_array.rb
