+++ week_1/2_difference_in_array.rb
