+++ week_1/2_no9.rb
