+++ week_1/2_string_concatenation_spec.rb
@@ -1,0 +1,10 @@
+require 'rspec'
+require '2_string_concatenation'
+
+RSpec.describe String_Concat do
+    let("abc")
+        it 'should concatenate two strings and remove any double characters' do
+            expect(conCatChar.new).to eql 
+        end
+    end
+
