+++ week_1/2_word_count.rb
