+++ week_1/gemfile
@@ -1,0 +1,2 @@
+source 'https://rubygems.org'
+gem 'rspec'
