--- week_1/helloworld.rb
@@ -1,1 +1,0 @@
-print "Hello, world!"
